Fix duplicated template parameter
The string "{0} jumps on to {0} and squishes them." on line 274 duplicates the
"{0}" template parameter, rendering the name of the command caller twice. This
corrects the second parameter to be "1" (the target user).

This commit fixes the error in the English localisation, but it exists in all
translations except for Swedish (it was either fixed or the template parameters
weren’t copied “correctly”). However, I’m not confident enough to touch the
French and Spanish translations and therefore I’ve opted to leave those as-is,
deferring it to those who actually speak the respective languages.